Small build fix

+++ build.sh
@@ -1,4 +1,7 @@
-rm flowaggs-server
+if [ -f flowaggs-server ]
+then
+  rm flowaggs-server
+fi
 cd cmd/flowaggs-server/
 go build
 cd ../../
